code/bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "A bot that implements all the steps to communicate with the telegram API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytelegrambotapi = "^4.14.0"
python-dotenv = "^1.0.0"
python-chess = "^1.999"
cairosvg = "^2.7.1"
asyncio = "^3.4.3"
aiohttp = "^3.9.1"
schedule = "^1.2.1"
aioschedule = "^0.5.2"
websocket-client = "^1.6.4"

[tool.poetry.group.test.dependencies]
coverage = "^7.3.2"

[tool.poetry.scripts]
start = 'bot:start'
tests = 'tests:tests'


[tool.coverage.run]
branch = true
source = ["bot"]
command_line = "-m unittest"

[tool.coverage.report]
show_missing = true

[tool.coverage.xml]
output = "coverage.xml"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
